sort layers, ensure ids are unique, paint context layers of their color, etc
