+++ libs/deck-layer-composer/src/index.ts
@@ -2,6 +2,5 @@ export * from './config'
 export * from './hooks/deck-layers.hooks'
 export * from './hooks/deck-layers-composer.hooks'
 export * from './hooks/deck-layers-state.hooks'
-export * from './deck-layer-composer'
 export * from './resolvers'
 export * from './utils'
