+++ libs/deck-layer-composer/src/resolvers/basemap.ts
@@ -1,10 +1,12 @@
 import { DataviewInstance } from '@globalfishingwatch/api-types'
 import { BaseMapLayerProps } from '@globalfishingwatch/deck-layers'
+import { GROUP_ORDER, Group } from '@globalfishingwatch/layer-composer'
 
 export function resolveDeckBasemapLayerProps(dataview: DataviewInstance): BaseMapLayerProps {
   return {
     id: dataview.id,
     visible: dataview.config?.visible || true,
     basemap: dataview.config?.basemap || 'default',
+    zIndex: GROUP_ORDER.indexOf(Group.Basemap),
   }
 }
