+++ libs/deck-layer-composer/src/resolvers/context.ts
@@ -1,5 +1,6 @@
 import { DataviewInstance } from '@globalfishingwatch/api-types'
-import { BaseMapLayerProps, ContextLayerProps } from '@globalfishingwatch/deck-layers'
+import { ContextLayerProps } from '@globalfishingwatch/deck-layers'
+import { GROUP_ORDER, Group } from '@globalfishingwatch/layer-composer'
 import { ResolverGlobalConfig } from '../resolvers/types'
 
 export function resolveDeckContextLayerProps(
@@ -10,6 +11,7 @@ export function resolveDeckContextLayerProps(
     id: dataview.id,
     color: dataview.config?.color!,
     datasetId: dataview.config?.layers?.[0].dataset!,
+    zIndex: GROUP_ORDER.indexOf(Group.OutlinePolygons),
     // hoveredFeatures,
     // clickedFeatures,
   }
